+++ R/check.R
@@ -6,7 +6,6 @@
 #' @param server RSC server. If NULL, use the ENV variable CONNECT_SERVER
 #' @param token RSC api token. If NULL, use the ENV variable CONNECT_API_KEY
 #' @param dir directory location of the the config file
-#' @param file config file name
 #' @param debug_level Integer, 0 to 2.
 #'
 #' @details
@@ -18,7 +17,7 @@
 #' @import uatBase
 #' @export
 check = function(server = NULL, token = NULL,
-                 dir = ".", file = "config-uat.yml",
+                 dir = ".",
                  debug_level = 0:2) {
 
   debug_level = get_debug_level(force(debug_level))
@@ -30,10 +29,10 @@ check = function(server = NULL, token = NULL,
   check_list$users_details = summarise_users(get_server(), get_token(), debug_level = debug_level)
   check_list$versions = summarise_versions(get_server(), get_token())
   check_list$sys_deps = check_sys_deps(debug_level = debug_level)
-  register_uat_user(get_server(), get_token(), account = check_list$user_details$username)
+  register_uat_user(get_server(), get_token(), account = get_account())
 
   cli::cli_h2("Document Deployment Checks")
-  r6_inits = init_r6_checks(dir = dir, file = file)
+  r6_inits = init_r6_checks(dir = dir, file = "config-uat-rsc.yml")
   lapply(r6_inits, function(r6) r6$check(debug_level = debug_level))
   check_list$deployments = purrr::map_dfr(r6_inits, ~.x$get_log())
   check_list$deployments = dplyr::arrange(check_list$deployments, .data$group, .data$short)